Convert the page/link chooser to use static onload handlers
