--- wagtail/admin/templates/wagtailadmin/chooser/external_link_chosen.js
@@ -1,4 +1,0 @@
-function(modal, jsonData) {
-    modal.respond('pageChosen', jsonData['result']);
-    modal.close();
-}
